pyproject.toml
[tool.poetry]
name = "autonom"
version = "0.0.1"
description = "The Ansible task runner for AWS IoT"
authors = ["rxnew"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/rxnew/autonom"

[tool.poetry.dependencies]
python = "^3.7"
awsiotpythonsdk = "^1.4"
docutils = "=0.15"
ansible-api = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
autonom = "autonom.__main__:main"
